+++ config/routes.rb
@@ -30,7 +30,11 @@ Rails.application.routes.draw do
       resources :budgets, only: [:new, :create, :show, :edit, :update]
     end
 
-    resources :activities, only: [], concerns: [:transactionable, :budgetable] do
+    concern :disbursement_plannable do
+      resources :planned_disbursements, only: [:new, :create]
+    end
+
+    resources :activities, only: [], concerns: [:transactionable, :budgetable, :disbursement_plannable] do
       resources :steps, controller: "activity_forms"
       resources :extending_organisations, only: [:edit, :update]
       resources :implementing_organisations, only: [:new, :create, :edit, :update]